+++ lib/braintree/transaction_gateway.rb
@@ -232,7 +232,7 @@ module Braintree
         ]},
         {:custom_fields => :_any_key_},
         {:descriptor => [:name, :phone, :url]},
-        {:paypal_account => [:email, :token, :paypal_data, :payee_id, :payee_email]},
+        {:paypal_account => [:email, :token, :paypal_data, :payee_id, :payee_email, :payer_id, :payment_id]},
         {:industry => [:industry_type, {:data => [:folio_number, :check_in_date, :check_out_date, :travel_package, :lodging_check_in_date, :lodging_check_out_date, :departure_date, :lodging_name, :room_rate]}]},
         {:apple_pay_card => [:number, :cardholder_name, :cryptogram, :expiration_month, :expiration_year, :eci_indicator]},
         {:android_pay_card => [:number, :cryptogram, :google_transaction_id, :expiration_month, :expiration_year, :source_card_type, :source_card_last_four, :eci_indicator]}