+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# unreleased
+* Allow PayPal payment ID and payer ID to be passed during transaction create
+
 # 2.92.0
 * Add `processor_response_type` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
 * Add `authorization_expires_at` to `Transaction`